Multiple Names and Yellow Underlines
Added functionality for entries with multiple names to be processed in
underlining. However, this feature seems to be a little bugged as MISON
and its 3 names are not being properly processed and underlined. Could
be a problem with capitilization or with storage of names in Hashset.

Underlines were changed to be yellow by swapping the order of
highlighting and underlining so that the highlight comes first and the
color tags are outside the underline tags making them yellow. Underline
was also moved from indent of 1 to 0 in text asset but doesn't seem too
different.

The required yellow clause was added with a regex pattern. Next up:
fixing mison and its glitch and adding saving. After that make it
optional in the settings for upcoming builds so that i can see with and
without it.